added: custom equality
Some functions that compare elements, such as Contains will use the following method if it is available on the element type:

    func (a ElementType) Equals(b ElementType) bool

The `ElementType` must be the same for the receiver and argument and it must return a bool. Be careful to create the function on the pointer or non-pointer
type that is used by the slice.

Here is a minimal example:

    type Car struct {
    	Name, Color string
    }

    type Cars []*Car // ElementType is *Car

    func (c *Car) Equals(c2 *Car) bool {
    	return c.Name == c2.Name
    }

Fixes #111