added: SequenceUsing (#122)
SequenceUsing generates slice in range using creator function

There are 3 variations to generate:

    1. [0, n).
    2. [min, max).
    3. [min, max) with step.

If len(params) == 1 considered that will be returned slice between 0 and n, where n is the first param, [0, n).

If len(params) == 2 considered that will be returned slice between min and max, where min is the first param, max is the second, [min, max).

If len(params) > 2 considered that will be returned slice between min and max with step, where min is the first param, max is the second, step is the third one, [min, max) with step, others params will be ignored.

Fixes #107