Added Equal() (#151)
Equals compare elements from the start to the end.

If they are the same is considered the slices are equal if all elements are the same is considered the slices are equal.

If each slice == nil is considered that they're equal.

If element realizes Equals interface it uses that method, in other way uses default compare.